+++ map/map_tests/map_foreach_test.cpp
@@ -221,7 +221,7 @@ namespace
 
         FeatureType::GetTypesFn getTypes;
         f.ForEachTypeRef(getTypes);
-        for (int i = 0; i < getTypes.m_size; ++i)
+        for (size_t i = 0; i < getTypes.m_size; ++i)
           cout << classif().GetFullObjectName(getTypes.m_types[i]) << endl;
       }
     }
